Using the Docker tests as a base code instead
