+++ test/multihoming-test/omsfiles/customlog.conf
@@ -4,7 +4,7 @@
   type sudo_tail
   path /var/log/custom.log
   pos_file /var/opt/microsoft/omsagent/state/CUSTOM_LOG_BLOB.Custom_Log_CL_<workspace-id>.pos
-  read_from_head true
+  read_from_head false
   run_interval 60
   tag oms.blob.CustomLog.CUSTOM_LOG_BLOB.Custom_Log_CL_<workspace-id>.*
   format none
