+++ test/multihoming-test/parameters.json
@@ -1,14 +1,14 @@
 {
   "resource": "https://management.azure.com",
+  "tenant": "72f988bf-86f1-41af-91ab-2d7cd011db47",
   "authority host url": "https://login.microsoftonline.com",
+  "app id": "<app-id>",
+  "app secret": "<app-secret>",
+  "oms bundle": "<bundle-file-name>",
+  "old oms bundle": "<old-bundle-file-name>",
   "resource group": "<resource-group-name>",
-  "location": "<location>",
-  "username": "<username>",
-  "ssh private": "<ssh-private-keyfile-path>",
-  "nsg resource group": "<nsg-resource-group>",
-  "nsg": "<nsg>",
-  "size": "<size>",
+  "subscription": "<subscription-id>",
   "workspace": "<workspace-name>",
-  "key vault": "<key-vault-name>",
-  "old version": "<old-extesion-version>"
+  "workspace id": "<workspace-id>",
+  "workspace key": "<workspace-key>"
 }
